refactor: move Inline.stories.tsx and Stack.stories.tsx to src/stories
